UPSTREAM: crypto: sha512-generic - move to generic glue implementation
This updated the generic SHA-512 implementation to use the
generic shared SHA-512 glue code.

It also implements a .finup hook crypto_sha512_finup() and exports
it to other modules. The import and export() functions and the
.statesize member are dropped, since the default implementation
is perfectly suitable for this module.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
(cherry picked from commit ca142584bc8ebc8e15b317680771ade58ca96315)
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: sweetyicecare <gabrielseedev@outlook.com>